Changed WaterHeater to SmartPlug. Refractor-ed the entire program with anything pertaining to WaterHeater.
