+++ src/main/java/seedu/smarthomebot/common/Messages.java
@@ -40,8 +40,6 @@ public class Messages {
     public static final String MESSAGE_ILLEGAL_CHARACTER = "Illegal Character space or / detected in";
     public static final String MESSAGE_EMPTY_PARAMETER =
             "Empty Parameter detected! Please follow format and enter require parameters.";
-    public static final String MESSAGE_INVALID_TEMPERATURE_WH =
-            "Invalid speed is inputted, ensure that it is within 20-50 degrees. \n "
-                    + "Previous set speed will be set.";
+
 
 }
